+++ CPP/BeginnerVillage/Scene/MainGame.tscn
@@ -19,7 +19,7 @@
 [ext_resource path="res://Scene/LogUIControl.gd" type="Script" id=17]
 [ext_resource path="res://Scene/EventLogPanelUI.gd" type="Script" id=18]
 
-[sub_resource type="GDScript" id=4]
+[sub_resource type="GDScript" id=1]
 script/source = "extends Node2D
 var input_manager
 
@@ -42,13 +42,13 @@ func draw_highlight_box(rect):
 	draw_rect(rect, Color.darkgray, false, 1.0, false)
 "
 
-[sub_resource type="DynamicFont" id=1]
-font_data = ExtResource( 4 )
-
 [sub_resource type="DynamicFont" id=2]
 font_data = ExtResource( 4 )
 
 [sub_resource type="DynamicFont" id=3]
+font_data = ExtResource( 4 )
+
+[sub_resource type="DynamicFont" id=4]
 font_data = ExtResource( 4 )
 
 [node name="Main" type="Node"]
@@ -84,7 +84,7 @@ __meta__ = {
 }
 
 [node name="InputEffect" type="Node2D" parent="GodotInput"]
-script = SubResource( 4 )
+script = SubResource( 1 )
 
 [node name="UIControl" type="Node" parent="."]
 script = ExtResource( 7 )
@@ -100,7 +100,7 @@ margin_left = 5.78651
 margin_top = 109.276
 margin_right = 121.787
 margin_bottom = 152.275
-custom_fonts/font = SubResource( 1 )
+custom_fonts/font = SubResource( 2 )
 text = "캐릭터 만들기"
 __meta__ = {
 "_edit_use_anchors_": false
@@ -138,7 +138,7 @@ margin_left = 61.7732
 margin_top = 133.576
 margin_right = 149.493
 margin_bottom = 123.776
-custom_fonts/font = SubResource( 2 )
+custom_fonts/font = SubResource( 3 )
 text = "test"
 align = 2
 __meta__ = {
@@ -152,7 +152,7 @@ margin_left = 966.667
 margin_top = 2.33334
 margin_right = 1021.67
 margin_bottom = 52.3333
-custom_fonts/font = SubResource( 3 )
+custom_fonts/font = SubResource( 4 )
 text = "가구 1"
 __meta__ = {
 "_edit_use_anchors_": false
