+++ CPP/NativeLib/NativeLib/NativeLib.cpp
@@ -1,11 +1,9 @@
 #include "Common.h"
-#include "TestManager.h"
 #include "GameManager.h"
 #include "UIManager.h"
 #include "WorldManager.h"
 #include "InputManager.h"
 #include "CameraManager.h"
-#include "PathfindingManager.h"
 
 extern "C" void GDN_EXPORT godot_gdnative_init(godot_gdnative_init_options * o) {
 	godot::Godot::gdnative_init(o);
@@ -18,11 +16,9 @@ extern "C" void GDN_EXPORT godot_gdnative_terminate(godot_gdnative_terminate_opt
 extern "C" void GDN_EXPORT godot_nativescript_init(void* handle) {
 	godot::Godot::nativescript_init(handle);
 
-	godot::register_class<TestManager>();
 	godot::register_class<GameManager>();
 	godot::register_class<UIManager>();
 	godot::register_class<WorldManager>();
 	godot::register_class<InputManager>();
 	godot::register_class<CameraManager>();
-	godot::register_class<PathfindingManager>();
 }
