--- CPP/NativeLib/NativeLib/PathfindingManager.h
@@ -1,29 +1,0 @@
-#pragma once
-#include "testProperty.h"
-#include "Base.h"
-#include <Node2D.hpp>
-
-class PathfindingManager : public Node2D {
-
-	GODOT_CLASS(PathfindingManager, Node2D);
-
-private:
-	shared_ptr <JPSCollision> spCollision = std::make_shared<JPSCollision>();
-
-public:
-	static void _register_methods();
-
-	int linktest = 3;
-	list<JPSCoord> ResultNodes;
-	PoolVector2Array ResultPaths;
-
-	void _init();
-	void _ready();
-	void InitMap(Int32 width, Int32 height);
-	void UpdateMap(Int32 gridWidth, Int32 gridHeight);
-	void ScanningCollision(Int32 halfWidth);
-
-	PoolVector2Array GetPath();
-	void Search(Int32 Sx, Int32 Sy, Int32 Ex, Int32 Ey);//, list<JPSCoord>& ResultNodes);
-	int GetPathLength() { return ResultNodes.size(); }
-};
